+++ main_project/component/diode.hpp
@@ -7,9 +7,9 @@ class Diode: public Component{
 private:
 	static constexpr float vt = 25e-3f; //Constants for now. Could determine from netlist later on
 	static constexpr float is = 1e-14f;
-	float compCurrent; //Value of the current source in the companion model
+	double compCurrent; //Value of the current source in the companion model
 	// float compVoltage; //value of the voltage across the diode
-	float compConductance; //Conductance of the resistor in the companion model
+	double compConductance; //Conductance of the resistor in the companion model
 
 public:
 	Diode(string name, vector<string> args, vector<float> extraInfo);
